S3/requirements.txt
boto3==1.26.32
pycryptodome==3.16.0
